pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sql-agent-llmops"
version = "0.1.0"
description = "Multi-model SQL Agent with RAG for natural language SQL generation and visualization"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "SQL Agent Team" }
]
keywords = [
    "sql",
    "llm",
    "rag",
    "natural-language",
    "visualization",
    "orchestration",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "transformers>=4.35.0",
    "torch>=2.0.0",
    "llama-cpp-python>=0.2.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "plotly>=5.17.0",
    "lxml>=4.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.7.0",
    "ruff>=0.0.285",
    "mypy>=1.5.0",
    "sphinx>=7.1.0",
]
llama = [
    "llama-cpp-python>=0.2.0",
]
visualization = [
    "plotly>=5.17.0",
    "kaleido>=0.2.1",
]

[project.urls]
Homepage = "https://github.com/yourusername/sql-agent-llmops"
Repository = "https://github.com/yourusername/sql-agent-llmops"
Documentation = "https://sql-agent-llmops.readthedocs.io"
Issues = "https://github.com/yourusername/sql-agent-llmops/issues"

[tool.setuptools]
packages = ["src"]

[tool.setuptools.package-data]
src = ["py.typed"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
]
ignore = [
    "E501",  # line too long
    "E741",  # ambiguous variable name
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
